fix: logging IsoFLOPs

+++ src/model/utils.py
@@ -339,7 +339,7 @@ def update_dynamic_cache(
     return dynamic_cache.key_cache[layer_idx], dynamic_cache.value_cache[layer_idx]
 
 
-def calculate_FLOPs(model: torch.nn.Module, tokenizer: ReLlamaTokenizer, max_seq_len: int) -> float:
+def calculate_FLOPs(model: torch.nn.Module, tokenizer: ReLlamaTokenizer, max_seq_len: int) -> int:
     flops, macs, params = calculate_flops(model=model, 
                                         input_shape=(1, max_seq_len),
                                         transformer_tokenizer=tokenizer,
@@ -347,4 +347,4 @@ def calculate_FLOPs(model: torch.nn.Module, tokenizer: ReLlamaTokenizer, max_seq
                                         print_results=False,
                                         print_detailed=False,
                                         output_as_string=False)
-    return flops
+    return int(flops)
